GP2-500: Extend render_video template tag to include duration as a special data attribute
To help get the watch time of a page featuring videos, and factoring in the requirement to skip personalised content, it seems easier to leverage the existing behaviour where the entire page is rendered post-save and its HTML is parsed to work out a readtime.

By putting the duration of the video in a custom data attribute, we can pull it out from the rendered HTML and factor it into our calculations.

Note that we're not using the 'duration' attribute that `<video>` can use, because that's a read-only attr that some browsers will set (but not all, it seems), so we want to stay well away from that to avoid a clash.

Also added some docstrings to help make clear what's going on with the simple video block vs the personalised one.