GP2-500: Add counting watching time as part of the overall page read time
This commit uses the existing parsed HTML of the rendered page to extract the data attributes we've added that show the readtime for any video found in the page. (For now there's only one possible, but that will grow in future changes).

Note that we have to do some fairly fiddly stuff to avoid either a) creating a new revision with the updated duration attached to it, meaning that the live page still shows an outdated duration and b) triggering save() which risks pushing draft content live, as Rafa spotted and fixed.

Includes tests for the core cases. Note that the main body StreamField only uses a Personalised Video block at the moment, and all personalised contnet is not in scope for page time calculations, so we're skipping that in this changeset.