Staging release (#402)
* EP optional fields

* Upgraded node-fetch package

* styling sign up and login flow

* remove airtable

* remove unused code

* fix browser tests

* Fix heading styling

* improve test coverage

* Minor update to the PR template

Mainly to make it easy to add a clear link to Jira, for reviewers' convenience, but also a couple of other nice-to-haves

* (GP2-449) Extended Video model for transcript text (#370)

* GP2-388_country_list

* Rebuild js

* bump version of directory constants

* Country type-ahead, nicer buttons and corrected headings

* Tests and changelog

* Rebuild assets

* rebuild assets

* Style fix

* GP2-594_GP2-595_Personalization_bar_buttons

* changelog

* style about your business page

* ConfirmationComponent for Product button

* GP2-500: Refactor out reusable code for making Video/Media entities into test helper

(This will be use in subsequent tests as part of this feature)

* GP2-500: Extend fake-video-creation helper to take extra parameters

...and handle multiple uses within the same test without clashing on Collection.objects.create

* GP2-500: Extend render_video template tag to include duration as a special data attribute

To help get the watch time of a page featuring videos, and factoring in the requirement to skip personalised content, it seems easier to leverage the existing behaviour where the entire page is rendered post-save and its HTML is parsed to work out a readtime.

By putting the duration of the video in a custom data attribute, we can pull it out from the rendered HTML and factor it into our calculations.

Note that we're not using the 'duration' attribute that `<video>` can use, because that's a read-only attr that some browsers will set (but not all, it seems), so we want to stay well away from that to avoid a clash.

Also added some docstrings to help make clear what's going on with the simple video block vs the personalised one.

* GP2-500: Add counting watching time as part of the overall page read time

This commit uses the existing parsed HTML of the rendered page to extract the data attributes we've added that show the readtime for any video found in the page. (For now there's only one possible, but that will grow in future changes).

Note that we have to do some fairly fiddly stuff to avoid either a) creating a new revision with the updated duration attached to it, meaning that the live page still shows an outdated duration and b) triggering save() which risks pushing draft content live, as Rafa spotted and fixed.

Includes tests for the core cases. Note that the main body StreamField only uses a Personalised Video block at the moment, and all personalised contnet is not in scope for page time calculations, so we're skipping that in this changeset.

* GP2-500: Update changelog

* GP2-500: Belated flake8 sweep

Have enabled flake8 over a lax pylint config now. Next time will be cleaner.

* GP2-500: Add additional test to confirm _update_data_for_appropriate_version operates as expected

This will be a handy regression test.

Note that the off-by-one result at the end of the test is unrelated to the function being tested.

* GP2-511 make sidebar collapsible

* sidebar variables

* fix tests

* navbar data

* styling and add units tests

* fix browser tests

* compiled assets

* GP2-260: Add a new custom block for a Choose/Do Not Choose element

All fields within the block are mandatory, so it's impossible to create the block without all the required fields.

The rich-text fields use a subset of the options for a standard rich-text field, limiting to OL, UL and line breaks

* GP2-260: Add new ChooseDoNotChoose block to DetailPage body StreamField

* GP2-260: Add initial/skeleton template for ChooseDoNotChoose block, ahead of proper FE implementation

Currently renders all the required fields, in rich-text but lacks the full design's elements and any proper styling.

* GP2-620: Remove support for lists (OL and UL) from the rich text for a Do Not Choose block

So for now all we support are line breaks

* Video transcript FE

* GP2-449: Update page readtime tests, following minor increase in page HTML volume

Also updated comments to help our future selves if these tests faily by off-by-one issues

* Add FE styling for choose dont choose block

* GP2-76: Backend changes to support a link from a module page back to its parent learning page

Includes a small test.

Needs FE work to complete this ticket - so over to LC.

* Update CSS assets

* Added next lesson to detail page (#360)

* Added missing image for next module in DetailPage (#378)

* Refactor

* Update pull_request_template.md

* saving of product-country details

* GP2-512 create input and textarea component

* Refactor

* Fix implementation after recent merge

* updated package-lock

* rename Index to index (#385)

* Add verification code modal

* Move verification code modal to signup template

* Add compiled dist files and edit CHANGELOG

* Build dist files

* Remove modal from base template

* Refactor

* Re enable css icons import

* Rebuild dist files

* Refactor

* Minor refactor to remove conditional

* GP2-657: Merge both templatetag test files into one

Still keeps them as a mix of pytest and unittest, but at least they are in a single place for discovery

* GP2-657: Add a template tag that allows adding a backlink to the source page to an outbound URL

eg: if a page with the path /my/source/page has a link to http://example.com/target-page it will turn that link to be  `http://example.com/target-page?return-link=%2Fmy%2Fsource%2Fpage`

The querystring name is defined as a constant (defaults to 'return-link') but this tag can be extended to use additional names in the future.

* GP2-657: Add tests for get_backlinked_url templatetag

* GP2-657: Add tests for DetailPage.get_context() to show backlink extraction working

* GP2-657: Add comments to clarify the _why_ of behaviour

* GP2-657: Update CHANGELOG.md

* GP2-657: flake8 fixups

* GP2-657: Update backlink extraction to ONLY accept backlinks for Export Plan, plus obtain the page title for the relevant EP page

Uses a constants map that already exists in the codebase

* Added country confirmation model

* fix educational component position and formgroup

* GP2-599_header_menu

* fix test run

* Added avatar image + user name

* remove default image

* Changed styling to blue

* Next lesson image fix (#390)

* update dashboard headers

* make country dynamic

* Restyle header and move local styles out of modal jsx

* Back to export plan link on lesson page

* rebuild css

* Remove old unused content from learning homepage

* Fix tests - add changelog

* Fix browser tests

* rebuild assets

* fix flake

* fix classname

* FE added for back link

* Response to comments

* update changelog

* Update FE assets

* sidebar and textarea padding fixes

* GP2-601_header_link_highlight

* Include tag styles

* Fixes for personalisation and export plan pages

* Remove icon left classes

* Add clock icon

* add learning

* update changelog

* Fix sidebar positioning

* style learn linking component

Co-authored-by: Rizwan Mansuri <mr.mansuri@gmail.com>
Co-authored-by: Rizwan Mansuri <rizwan.mansuri@digital.trade.gov.uk>
Co-authored-by: Gerald Isabelle <info@geraldisabelle.me>
Co-authored-by: Lewis Coulson <66948936+lewis-coulson-dit@users.noreply.github.com>
Co-authored-by: Lewis Coulson <lewis@digital.trade.gov.uk>
Co-authored-by: Steve Jalim <steve@somefantastic.co.uk>
Co-authored-by: Alex Giamas <alexandros.giamas@digital.trade.gov.uk>
Co-authored-by: Bob Meredith <bob.meredith@ntlworld.com>
Co-authored-by: Rafa Garcia <rafa@rafagarcia.info>