GP2-667: Add a function to provide a limited, environment-appropriate configuration for Wagtail Transfer
This changeset adds a function, `get_wagtail_transfer_configuration`, that provides the appropriate Wagtail-Transfer config for the environment the app is running on.

Its current scope is limited to to Beta, Staging, Dev or local development. We will add Production when the time is right.

We don't want to have a single config that basically allows and deployed site to import from any other deployed site, because this opens the door to content from Dev being imported to Beta (or Staging and then on to Beta) in error.

Includes tests.

Before this is merged the env vars for the deployed instances will need updating to ensure they have the appropriate values set:

Beta needs to know credentials for the single site it can pull from:
    WAGTAILTRANSFER_BASE_URL_STAGING
    WAGTAILTRANSFER_SECRET_KEY_STAGING

Staging needs to know credentials for the single site it can pull from:
    WAGTAILTRANSFER_BASE_URL_BETA
    WAGTAILTRANSFER_SECRET_KEY_BETA

Dev needs to know credentials for the TWO sites it can pull from:
    WAGTAILTRANSFER_BASE_URL_STAGING
    WAGTAILTRANSFER_SECRET_KEY_STAGING
    WAGTAILTRANSFER_BASE_URL_BETA
    WAGTAILTRANSFER_SECRET_KEY_BETA