GP2-579: Make Redirects appear as Snippets, so we can transfer them with W-T
By default, wagtail.contrib.redirects.Redirect is not ported over, even if a page which features it is.

Note that you will need to port the target page first if you need to move a redirect. Even if the page has a new PK on the destination side, the Redirect WILL (i've checked) point to the correct page after transfer.