GP2-726: Redirect unauthenticated users on CMS pages to the sign-in page
This changeset adds`mixins.AuthenticatedUserRequired` to the base page used for most CMS pages (`CMSGenericPage`) so that, if the user is unauthenticated, they are redirected to the login page.

This also has the (requested/desired) side effect that the existing behaviour upon logout (kill session and reload the same page) now causes a redirect when the page is reloaded. Ideally we'd skip the reload, but given this is is a temporary change, I think it's fine for now)

Note that when a user is redirected to signin, we don't remember where they wanted to go originally. This is not ideal, but tolerable within the AC for this piece.