GP2-164: Wire up topic-level progress counter
Note the extra check that the completed-lessons data has to be a subset of the list of lessons available in the topic. This should be the case, because the data is pre-processed, but this check (and the decision not to show a counter if the data is bad) is there just in case.

# Conflicts:
#	core/templatetags/content_tags.py
#	tests/unit/core/test_templatetags.py