GP2-962: Add data migration to re-parent the lesson pages in the system
For every CuratedListPage, add 1...n TopicPages as their direct children, then make
the appropriate new TopicPage as the parent of the existing DetailPage (lesson)
and/or create a LessonPlaceholderPage as needed.

The source of which lesson to put on which topic comes from the CuratedListPage.topics StreamField

Includes a test of the reparenting function.