GP2-884: Defend against lessons that aren't in topics
At this point in time, it's STILL possible for a DetailPage to be the direct child of a CuratedListPage and NOT configured in a topic, so the data migration 0044 will not have moved it.  As a result, the previous version of `get_all_lessons` was returning all lessons regardless of topic configuration, which would be too many/ones we don't want.

This change gives get_all_lessons a narrower scope.

It's a shame not to keep the check added here happening at the DB level - maybe we can improve it later