GP2-652: Handle expired staff SSO token
This should, I hope, stop the issue where Wagtail 500s when faced with an expired SSO session by re-triggering the SSO auth flow.

It makes sure we only send people to the relevant admin login (Wagtail or Django) if we know were trying to go there.

(It MAY be we need to also zap the cookies, if this isn't enough.)