GP2-1144: Allow Beta token checks to support sub-paths of allowed paths
This changeset should reduce the noise in Sentry for paths which lack a Beta token but don't need the token anyway.

Plus, this fixes a broken UI issue on Beta:  `/static/` is currently configured as an allowed path, but only as a distinct, _complete_ path, rather than the head of a longer path to some static assets. This means that the login/landing page for a user on Beta who does NOT have a token lacks all the JS and CSS we'd want them to see. This bug should be fixed by these changes.

Adds tests and tweaks the BETA_WHITELISTED_ENDPOINT default vals for local dev (commented out) and test (used by pytest)