GP2-1284: Port the CountryGuidePage and dependents from BAU
Wagtail/Django level only.

Also brings ArticlePage, CampaignArticlePage and ArticleListingPage, whcih are all possible keyed/child-page types of CountryGuidePage

Templates and directory-components widgets will follow, but this gets the main modelling in.

It's VERY repetitive at the moment, because this single piece of work is a lift and shift - we may well rationalise the code in a subsequent ticket