GP2-1363: Drop custom Django view serving /markets/
This changeset removes the Django view serving /markets/ because this URL path is required for pages that will be ported from Great V1.

The view is not being moved to respond to a different route - it is no longer required, so it is safe to remove.

Related tests and template were removed.

Other tests that used /markets/ as an example URL have been updated so that they still pass