GP2-1190: Remove wagtail-personalisation - step 1 of n
Drop _usage_ of wagtail-personalisation within current code, including removing its rules and its streamfield blocks

Note that removing the PersonalisedStructBlock does not result in the loss of the StreamField content that it wrapped - it remains in the page and can be used as normal. Any content targetted at a specific segment, however, will become shown for all users.