GP2-1619: Enusure all pages ported from V1 into Domestic support DataLayer
This changeset makes the inclusion of the two GA360-specific mixins easier
by wrapping them into their own abstract mixin class.

This new DataLayerMixin is then added to the BaseContentPage for ex-V1 content or
directly into any page that does not use BaseContentPage.

There is also a test that checks that all concrete subclasses of wagtail.Page
in domestic.models implement the DataLayerMixin