GP2-1686: Sketch in ContactSuccessSnippet + grouped modeladmin
We now have a place where we can define editable content for the success pages in the contact form.

Note that for now, all the relevant sucess/topic options are available, to save having to add a new migration for each one we add later as we port over behaviour from V1

Tests to come.