GP2-1483: Squash migrations core.0001-0038 and omit references to wagtail_personalisation
Note that models related to W-P rules that were created had to be commented out in this squashd migration to remove our dependency on using ANY code from wagtail_personalisation.

However, because those models were later deleted in core.0053 (which has run on all environments), 0053 had to be retrospectively amended, to skip removal of models and relations which (now) are no longer being added.