GP2-1900: Fixup disabled tests by temporarily switching them to hard-coded paths
While we're running in hybrid V2 + V1-for-unmigrated-paths mode, it would be impossible to make the disabled tests pass, because some paths don't (and won't) exist to be `reverse()`d. As such, I've swapped that function call for a hard-coded string which would be its output. This does mean we're a bit more brittle if we move URLs in the future, but that's currently very unlikely and we can revert to looking up paths based on reverse() once V1 is entirely folded into V2

Also flattened out the feature flag ported from V1 -- V2 doesn't nest them in a dictionary, and adding a lookup dict for just a single flag now seemed to complicate matters