GP2-1616: Refactor Trade Finance to be a full wagtail page, not CMS snippet + Django view
This changeset reimplements how content for the /trade-finance/ is managed. V1 had it as a Django view + CMS-sourced content, so we initially went with the 'snippet' approach from the contact app. However literally all of the page's content went into that snippet, so we decided to refactor it to a simple (single-use) Wagtail page

Changes:
* refactor TradeFinanceSnippet to TradeFinancePage, incl migration that drops the old model + page heirarchy tests
* remove TradeFinanceSnippet from the modeladmin config (leaving case studies intact)
* rewire trade_finance.html template to expect a `page` object not `content_snippet`, plus tweak some url lookups to be slug-based not regular Django URLS
* drop now-redundant domestic.finance.views and its related tests