GP2-1850: Ensure pull quotes in ArticlePage.article_body are send to search index
Based on what's done in V1, literally every string of content in the MarkdownField used to make the article and its pull quotes is included, so this covers URLs as well.

It may be that we tune this detail out in the near future, but for now, parity is the goal.