GP2-2240: Improve stability of search results ordering/boosting
We've been seeing some unexpected/flip-flopping ordering of results on Dev and can't fully explain why the ranking scores for Articles vs Services

This changeset contains a recommendation from @michalc (the creator of DIT's ActivityStream) which moves the boost aspect of the query away from a 'should' clause with potential full-text search matching on the type (which may or may not be conflicting with the un-namespaced Article-type and Service-type records also being ingested from BAU into AS). Instead, we're trying out some clear function-based boosting.

This appears to be better:
1. The scores we're seeing for items returned correspond more predictably to the boosting involved (eg before Services were boosted by 20000, but the score was rarely higher than 13100 for a top-ranked result. Now, the boosted rank for a given Service document is literally the boost factor * rank of the Service document, so tuning is much easier to see.

2. We've maintained a similar 'watershed' between Services and Articles or Markets, but reduced the impact by an order of magnitiude. This is feeling like something of an art, but so far local tests seem promising