GP2-2155: Add healthcheck views
This changeset brings over /healthcheck/ and /healthcheck/ping/ from V1  to use in Magna

They use the same logic/checks of the required backends (API, Forms, SSO, Search).

The Search backend has been minorly updated to match the updated test data available only on the dev env, so the healthcheck only checks search in dev. This is the same as for V1