GP2-2721: Add support for subtitles on GreatMedia videos
This changeset adds field to GreatMedia for English subtitles (for now - more to come) which are then
served to the frontend via a dedicated view. The regular HTML5 video player picks up these subtitles
via a dedicated <track> node per language.

Note that VTT subtitles are required, and if the subtitle format is at all incorrect, things will not load.
This is partly why the subtitles are stored flat in a textfield, rather than being in a separate file: it's
much easier to tweak/tune them this way.

By default, no subtitles are shown - they have to be manually enabled.