GP2-2816 - Set Wagtaildocs serve method to 'direct' to avoid denial when proxying at PaaS level
Prior to this changeset, attempts to view docs on deployed instances were being proxied by Wagtail
(this is the default behviour). However, something at the permissions level was denying this, so
we've quickly switched to just providing the direct-access URL for documents. The downside of this
is we lose metrics/logging and also access control, but all of the files are public - there are are
no private documents in the project.