GP2-1683: Update defaults for some contact URLs to match production defaults
CAPITAL_INVEST_CONTACT_URL and FIND_A_SUPPLIER_CONTACT_URL will work across all envs if they are set to relative URLs as their default paths.
(This was already the case with FIND_A_SUPPLIER_CONTACT_URL in production - it used a relative path)

This simplifies what needs to go into the env vars (or, rather, removes the need to add as many things), but still lets them be configurable if we need it.