GP2-2988: BUGFIX: Make the name fields in the SOO contact wizard to still be editable if the first and/or last name is not yet known
This is needed because Great V2 allows users to sign up without providing a first or last name, which would then cause the overall wizard to 500.

NOTE THAT this fix is being applied to the pages after they have been migrated from V1 into V2, but if we decide we want to hotfix this to production on GDUI, it is a copy-paste change to port over.

Tests work at the Python level, but tested manually, too, of course.