GP2-3087: Main CH signup flow working
This involved a little backfilling of namespaces on redirect() calls, adding in a missing d-comp element and most of all changing the company lookup calls:

While d-components presents a dit object with dit.lookup.Foo, these were moved to GOVUK.Foo in great-domestic-ui, and so when that code was ported to great-cms, GOVUK.Foo was the namespace used in the ported company-lookup.js file. Updating the calls to that module from the enrolment templates seemed to be the cleanest solution, in terms of overall consistency of USE, even if it's not consistency of naming. Ideally we'd standardise this all back to use the dit.components namespace, but literally overwriting the current one (from GDUI) with the latest version in d-comp isn't viable as the one from GDUI (which uses GOVUK namespace) is ahead of / has diverged from the one in d-components, presumably for good reason - and this happened before the V1 to V2 migration.