feat: only include non-expired export opportunities in search
The Activity Stream currently only includes non-expired opportunities,
so when this goes live, there will not be any change of user-facing
behaviour.

However, the Activity Stream will soon be filled with expired
opportunities, so this commit is to make sure that the user-facing
behaviour will not change when that happens.