Reorder steps in Postgres init script
Change the order of the Postgres init script so that the required
roles and databases are configured before trying to populate the
database. This is because the population step can fail which would
prevent the roles and databases from being created on init. If the
population fail, the database can be populated manually afterwards.